+++ hexstrike_mcp.py
@@ -291,904 +291,18 @@ def setup_mcp_server(hexstrike_client: HexStrikeClient, compact: bool = False) -
     from mcp_tools.process_management import register_process_management_tools
     register_process_management_tools(mcp, hexstrike_client, logger)
 
-    # ============================================================================
-    # ADVANCED VULNERABILITY INTELLIGENCE MCP TOOLS (v6.0 ENHANCEMENT)
-    # ============================================================================
-
-    @mcp.tool()
-    def monitor_cve_feeds(hours: int = 24, severity_filter: str = "HIGH,CRITICAL", keywords: str = "") -> Dict[str, Any]:
-        """
-        Monitor CVE databases for new vulnerabilities with AI analysis.
-
-        Args:
-            hours: Hours to look back for new CVEs (default: 24)
-            severity_filter: Filter by CVSS severity - comma-separated values (LOW,MEDIUM,HIGH,CRITICAL,ALL)
-            keywords: Filter CVEs by keywords in description (comma-separated)
-
-        Returns:
-            Latest CVEs with exploitability analysis and threat intelligence
-
-        Example:
-            monitor_cve_feeds(48, "CRITICAL", "remote code execution")
-        """
-        data = {
-            "hours": hours,
-            "severity_filter": severity_filter,
-            "keywords": keywords
-        }
-        logger.info(f"🔍 Monitoring CVE feeds for last {hours} hours | Severity: {severity_filter}")
-        result = hexstrike_client.safe_post("api/vuln-intel/cve-monitor", data)
-
-        if result.get("success"):
-            cve_count = len(result.get("cve_monitoring", {}).get("cves", []))
-            exploit_analysis_count = len(result.get("exploitability_analysis", []))
-            logger.info(f"✅ Found {cve_count} CVEs with {exploit_analysis_count} exploitability analyses")
-
-        return result
-
-    @mcp.tool()
-    def generate_exploit_from_cve(cve_id: str, target_os: str = "", target_arch: str = "x64", exploit_type: str = "poc", evasion_level: str = "none") -> Dict[str, Any]:
-        """
-        Generate working exploits from CVE information using AI-powered analysis.
-
-        Args:
-            cve_id: CVE identifier (e.g., CVE-2024-1234)
-            target_os: Target operating system (windows, linux, macos, any)
-            target_arch: Target architecture (x86, x64, arm, any)
-            exploit_type: Type of exploit to generate (poc, weaponized, stealth)
-            evasion_level: Evasion sophistication (none, basic, advanced)
-
-        Returns:
-            Generated exploit code with testing instructions and evasion techniques
-
-        Example:
-            generate_exploit_from_cve("CVE-2024-1234", "linux", "x64", "weaponized", "advanced")
-        """
-        data = {
-            "cve_id": cve_id,
-            "target_os": target_os,
-            "target_arch": target_arch,
-            "exploit_type": exploit_type,
-            "evasion_level": evasion_level
-        }
-        logger.info(f"🤖 Generating {exploit_type} exploit for {cve_id} | Target: {target_os} {target_arch}")
-        result = hexstrike_client.safe_post("api/vuln-intel/exploit-generate", data)
-
-        if result.get("success"):
-            cve_analysis = result.get("cve_analysis", {})
-            exploit_gen = result.get("exploit_generation", {})
-            exploitability = cve_analysis.get("exploitability_level", "UNKNOWN")
-            exploit_success = exploit_gen.get("success", False)
-
-            logger.info(f"📊 CVE Analysis: {exploitability} exploitability")
-            logger.info(f"🎯 Exploit Generation: {'SUCCESS' if exploit_success else 'FAILED'}")
-
-        return result
-
-    @mcp.tool()
-    def discover_attack_chains(target_software: str, attack_depth: int = 3, include_zero_days: bool = False) -> Dict[str, Any]:
-        """
-        Discover multi-stage attack chains for target software with vulnerability correlation.
-
-        Args:
-            target_software: Target software/system (e.g., "Apache HTTP Server", "Windows Server 2019")
-            attack_depth: Maximum number of stages in attack chain (1-5)
-            include_zero_days: Include potential zero-day vulnerabilities in analysis
-
-        Returns:
-            Attack chains with vulnerability combinations, success probabilities, and exploit availability
-
-        Example:
-            discover_attack_chains("Apache HTTP Server 2.4", 4, True)
-        """
-        data = {
-            "target_software": target_software,
-            "attack_depth": min(max(attack_depth, 1), 5),  # Clamp between 1-5
-            "include_zero_days": include_zero_days
-        }
-        logger.info(f"🔗 Discovering attack chains for {target_software} | Depth: {attack_depth} | Zero-days: {include_zero_days}")
-        result = hexstrike_client.safe_post("api/vuln-intel/attack-chains", data)
-
-        if result.get("success"):
-            chains = result.get("attack_chain_discovery", {}).get("attack_chains", [])
-            enhanced_chains = result.get("attack_chain_discovery", {}).get("enhanced_chains", [])
-
-            logger.info(f"📊 Found {len(chains)} attack chains")
-            if enhanced_chains:
-                logger.info(f"🎯 Enhanced {len(enhanced_chains)} chains with exploit analysis")
-
-        return result
-
-    @mcp.tool()
-    def research_zero_day_opportunities(target_software: str, analysis_depth: str = "standard", source_code_url: str = "") -> Dict[str, Any]:
-        """
-        Automated zero-day vulnerability research using AI analysis and pattern recognition.
-
-        Args:
-            target_software: Software to research for vulnerabilities (e.g., "nginx", "OpenSSL")
-            analysis_depth: Depth of analysis (quick, standard, comprehensive)
-            source_code_url: URL to source code repository for enhanced analysis
-
-        Returns:
-            Potential vulnerability areas with exploitation feasibility and research recommendations
-
-        Example:
-            research_zero_day_opportunities("nginx 1.20", "comprehensive", "https://github.com/nginx/nginx")
-        """
-        if analysis_depth not in ["quick", "standard", "comprehensive"]:
-            analysis_depth = "standard"
-
-        data = {
-            "target_software": target_software,
-            "analysis_depth": analysis_depth,
-            "source_code_url": source_code_url
-        }
-        logger.info(f"🔬 Researching zero-day opportunities in {target_software} | Depth: {analysis_depth}")
-        result = hexstrike_client.safe_post("api/vuln-intel/zero-day-research", data)
-
-        if result.get("success"):
-            research = result.get("zero_day_research", {})
-            potential_vulns = len(research.get("potential_vulnerabilities", []))
-            risk_score = research.get("risk_assessment", {}).get("risk_score", 0)
-
-            logger.info(f"📊 Found {potential_vulns} potential vulnerability areas")
-            logger.info(f"🎯 Risk Score: {risk_score}/100")
-
-        return result
-
-    @mcp.tool()
-    def correlate_threat_intelligence(indicators: str, timeframe: str = "30d", sources: str = "all") -> Dict[str, Any]:
-        """
-        Correlate threat intelligence across multiple sources with advanced analysis.
-
-        Args:
-            indicators: Comma-separated IOCs (IPs, domains, hashes, CVEs, etc.)
-            timeframe: Time window for correlation (7d, 30d, 90d, 1y)
-            sources: Intelligence sources to query (cve, exploit-db, github, twitter, all)
-
-        Returns:
-            Correlated threat intelligence with attribution, timeline, and threat scoring
-
-        Example:
-            correlate_threat_intelligence("CVE-2024-1234,192.168.1.100,malware.exe", "90d", "all")
-        """
-        # Validate timeframe
-        valid_timeframes = ["7d", "30d", "90d", "1y"]
-        if timeframe not in valid_timeframes:
-            timeframe = "30d"
-
-        # Parse indicators
-        indicator_list = [i.strip() for i in indicators.split(",") if i.strip()]
-
-        if not indicator_list:
-            logger.error("❌ No valid indicators provided")
-            return {"success": False, "error": "No valid indicators provided"}
-
-        data = {
-            "indicators": indicator_list,
-            "timeframe": timeframe,
-            "sources": sources
-        }
-        logger.info(f"🧠 Correlating threat intelligence for {len(indicator_list)} indicators | Timeframe: {timeframe}")
-        result = hexstrike_client.safe_post("api/vuln-intel/threat-feeds", data)
-
-        if result.get("success"):
-            threat_intel = result.get("threat_intelligence", {})
-            correlations = len(threat_intel.get("correlations", []))
-            threat_score = threat_intel.get("threat_score", 0)
-
-            logger.info(f"📊 Found {correlations} threat correlations")
-            logger.info(f"🎯 Overall Threat Score: {threat_score:.1f}/100")
-
-        return result
-
-    @mcp.tool()
-    def advanced_payload_generation(attack_type: str, target_context: str = "", evasion_level: str = "standard", custom_constraints: str = "") -> Dict[str, Any]:
-        """
-        Generate advanced payloads with AI-powered evasion techniques and contextual adaptation.
-
-        Args:
-            attack_type: Type of attack (rce, privilege_escalation, persistence, exfiltration, xss, sqli)
-            target_context: Target environment details (OS, software versions, security controls)
-            evasion_level: Evasion sophistication (basic, standard, advanced, nation-state)
-            custom_constraints: Custom payload constraints (size limits, character restrictions, etc.)
-
-        Returns:
-            Advanced payloads with multiple evasion techniques and deployment instructions
-
-        Example:
-            advanced_payload_generation("rce", "Windows 11 + Defender + AppLocker", "nation-state", "max_size:256,no_quotes")
-        """
-        valid_attack_types = ["rce", "privilege_escalation", "persistence", "exfiltration", "xss", "sqli", "lfi", "ssrf"]
-        valid_evasion_levels = ["basic", "standard", "advanced", "nation-state"]
-
-        if attack_type not in valid_attack_types:
-            attack_type = "rce"
-
-        if evasion_level not in valid_evasion_levels:
-            evasion_level = "standard"
-
-        data = {
-            "attack_type": attack_type,
-            "target_context": target_context,
-            "evasion_level": evasion_level,
-            "custom_constraints": custom_constraints
-        }
-        logger.info(f"🎯 Generating advanced {attack_type} payload | Evasion: {evasion_level}")
-        if target_context:
-            logger.info(f"🎯 Target Context: {target_context}")
-
-        result = hexstrike_client.safe_post("api/ai/advanced-payload-generation", data)
-
-        if result.get("success"):
-            payload_gen = result.get("advanced_payload_generation", {})
-            payload_count = payload_gen.get("payload_count", 0)
-            evasion_applied = payload_gen.get("evasion_level", "none")
-
-            logger.info(f"📊 Generated {payload_count} advanced payloads")
-            logger.info(f"🛡️ Evasion Level Applied: {evasion_applied}")
-
-        return result
-
-    @mcp.tool()
-    def vulnerability_intelligence_dashboard() -> Dict[str, Any]:
-        """
-        Get a comprehensive vulnerability intelligence dashboard with latest threats and trends.
-
-        Returns:
-            Dashboard with latest CVEs, trending vulnerabilities, exploit availability, and threat landscape
-
-        Example:
-            vulnerability_intelligence_dashboard()
-        """
-        logger.info("📊 Generating vulnerability intelligence dashboard")
-
-        # Get latest critical CVEs
-        latest_cves = hexstrike_client.safe_post("api/vuln-intel/cve-monitor", {
-            "hours": 24,
-            "severity_filter": "CRITICAL",
-            "keywords": ""
-        })
-
-        # Get trending attack types
-        trending_research = hexstrike_client.safe_post("api/vuln-intel/zero-day-research", {
-            "target_software": "web applications",
-            "analysis_depth": "quick"
-        })
-
-        # Compile dashboard
-        dashboard = {
-            "timestamp": time.time(),
-            "latest_critical_cves": latest_cves.get("cve_monitoring", {}).get("cves", [])[:5],
-            "threat_landscape": {
-                "high_risk_software": ["Apache HTTP Server", "Microsoft Exchange", "VMware vCenter", "Fortinet FortiOS"],
-                "trending_attack_vectors": ["Supply chain attacks", "Cloud misconfigurations", "Zero-day exploits", "AI-powered attacks"],
-                "active_threat_groups": ["APT29", "Lazarus Group", "FIN7", "REvil"],
-            },
-            "exploit_intelligence": {
-                "new_public_exploits": "Simulated data - check exploit-db for real data",
-                "weaponized_exploits": "Monitor threat intelligence feeds",
-                "exploit_kits": "Track underground markets"
-            },
-            "recommendations": [
-                "Prioritize patching for critical CVEs discovered in last 24h",
-                "Monitor for zero-day activity in trending attack vectors",
-                "Implement advanced threat detection for active threat groups",
-                "Review security controls against nation-state level attacks"
-            ]
-        }
-
-        logger.info("✅ Vulnerability intelligence dashboard generated")
-        return {
-            "success": True,
-            "dashboard": dashboard
-        }
-
-    @mcp.tool()
-    def threat_hunting_assistant(target_environment: str, threat_indicators: str = "", hunt_focus: str = "general") -> Dict[str, Any]:
-        """
-        AI-powered threat hunting assistant with vulnerability correlation and attack simulation.
-
-        Args:
-            target_environment: Environment to hunt in (e.g., "Windows Domain", "Cloud Infrastructure")
-            threat_indicators: Known IOCs or suspicious indicators to investigate
-            hunt_focus: Focus area (general, apt, ransomware, insider_threat, supply_chain)
-
-        Returns:
-            Threat hunting playbook with detection queries, IOCs, and investigation steps
-
-        Example:
-            threat_hunting_assistant("Windows Domain", "suspicious_process.exe,192.168.1.100", "apt")
-        """
-        valid_hunt_focus = ["general", "apt", "ransomware", "insider_threat", "supply_chain"]
-        if hunt_focus not in valid_hunt_focus:
-            hunt_focus = "general"
-
-        logger.info(f"🔍 Generating threat hunting playbook for {target_environment} | Focus: {hunt_focus}")
-
-        # Parse indicators if provided
-        indicators = [i.strip() for i in threat_indicators.split(",") if i.strip()] if threat_indicators else []
-
-        # Generate hunting playbook
-        hunting_playbook = {
-            "target_environment": target_environment,
-            "hunt_focus": hunt_focus,
-            "indicators_analyzed": indicators,
-            "detection_queries": [],
-            "investigation_steps": [],
-            "threat_scenarios": [],
-            "mitigation_strategies": []
-        }
-
-        # Environment-specific detection queries
-        if "windows" in target_environment.lower():
-            hunting_playbook["detection_queries"] = [
-                "Get-WinEvent | Where-Object {$_.Id -eq 4688 -and $_.Message -like '*suspicious*'}",
-                "Get-Process | Where-Object {$_.ProcessName -notin @('explorer.exe', 'svchost.exe')}",
-                "Get-ItemProperty HKLM:\\Software\\Microsoft\\Windows\\CurrentVersion\\Run",
-                "Get-NetTCPConnection | Where-Object {$_.State -eq 'Established' -and $_.RemoteAddress -notlike '10.*'}"
-            ]
-        elif "cloud" in target_environment.lower():
-            hunting_playbook["detection_queries"] = [
-                "CloudTrail logs for unusual API calls",
-                "Failed authentication attempts from unknown IPs",
-                "Privilege escalation events",
-                "Data exfiltration indicators"
-            ]
-
-        # Focus-specific threat scenarios
-        focus_scenarios = {
-            "apt": [
-                "Spear phishing with weaponized documents",
-                "Living-off-the-land techniques",
-                "Lateral movement via stolen credentials",
-                "Data staging and exfiltration"
-            ],
-            "ransomware": [
-                "Initial access via RDP/VPN",
-                "Privilege escalation and persistence",
-                "Shadow copy deletion",
-                "Encryption and ransom note deployment"
-            ],
-            "insider_threat": [
-                "Unusual data access patterns",
-                "After-hours activity",
-                "Large data downloads",
-                "Access to sensitive systems"
-            ]
-        }
-
-        hunting_playbook["threat_scenarios"] = focus_scenarios.get(hunt_focus, [
-            "Unauthorized access attempts",
-            "Suspicious process execution",
-            "Network anomalies",
-            "Data access violations"
-        ])
-
-        # Investigation steps
-        hunting_playbook["investigation_steps"] = [
-            "1. Validate initial indicators and expand IOC list",
-            "2. Run detection queries and analyze results",
-            "3. Correlate events across multiple data sources",
-            "4. Identify affected systems and user accounts",
-            "5. Assess scope and impact of potential compromise",
-            "6. Implement containment measures if threat confirmed",
-            "7. Document findings and update detection rules"
-        ]
-
-        # Correlate with vulnerability intelligence if indicators provided
-        if indicators:
-            logger.info(f"🧠 Correlating {len(indicators)} indicators with threat intelligence")
-            correlation_result = correlate_threat_intelligence(",".join(indicators), "30d", "all")
-
-            if correlation_result.get("success"):
-                hunting_playbook["threat_correlation"] = correlation_result.get("threat_intelligence", {})
-
-        logger.info("✅ Threat hunting playbook generated")
-        return {
-            "success": True,
-            "hunting_playbook": hunting_playbook
-        }
-
-    # ============================================================================
-    # ENHANCED VISUAL OUTPUT TOOLS
-    # ============================================================================
-
-    @mcp.tool()
-    def get_live_dashboard() -> Dict[str, Any]:
-        """
-        Get a beautiful live dashboard showing all active processes with enhanced visual formatting.
-
-        Returns:
-            Live dashboard with visual process monitoring and system metrics
-        """
-        logger.info("📊 Fetching live process dashboard")
-        result = hexstrike_client.safe_get("api/processes/dashboard")
-        if result.get("success", True):
-            logger.info("✅ Live dashboard retrieved successfully")
-        else:
-            logger.error("❌ Failed to retrieve live dashboard")
-        return result
-
-    @mcp.tool()
-    def create_vulnerability_report(vulnerabilities: str, target: str = "", scan_type: str = "comprehensive") -> Dict[str, Any]:
-        """
-        Create a beautiful vulnerability report with severity-based styling and visual indicators.
-
-        Args:
-            vulnerabilities: JSON string containing vulnerability data
-            target: Target that was scanned
-            scan_type: Type of scan performed
-
-        Returns:
-            Formatted vulnerability report with visual enhancements
-        """
-        import json
-
-        try:
-            # Parse vulnerabilities if provided as JSON string
-            if isinstance(vulnerabilities, str):
-                vuln_data = json.loads(vulnerabilities)
-            else:
-                vuln_data = vulnerabilities
-
-            logger.info(f"📋 Creating vulnerability report for {len(vuln_data)} findings")
-
-            # Create individual vulnerability cards
-            vulnerability_cards = []
-            for vuln in vuln_data:
-                card_result = hexstrike_client.safe_post("api/visual/vulnerability-card", vuln)
-                if card_result.get("success"):
-                    vulnerability_cards.append(card_result.get("vulnerability_card", ""))
-
-            # Create summary report
-            summary_data = {
-                "target": target,
-                "vulnerabilities": vuln_data,
-                "tools_used": [scan_type],
-                "execution_time": 0
-            }
-
-            summary_result = hexstrike_client.safe_post("api/visual/summary-report", summary_data)
-
-            logger.info("✅ Vulnerability report created successfully")
-            return {
-                "success": True,
-                "vulnerability_cards": vulnerability_cards,
-                "summary_report": summary_result.get("summary_report", ""),
-                "total_vulnerabilities": len(vuln_data),
-                "timestamp": summary_result.get("timestamp", "")
-            }
-
-        except Exception as e:
-            logger.error(f"❌ Failed to create vulnerability report: {str(e)}")
-            return {"success": False, "error": str(e)}
-
-    @mcp.tool()
-    def format_tool_output_visual(tool_name: str, output: str, success: bool = True) -> Dict[str, Any]:
-        """
-        Format tool output with beautiful visual styling, syntax highlighting, and structure.
-
-        Args:
-            tool_name: Name of the security tool
-            output: Raw output from the tool
-            success: Whether the tool execution was successful
-
-        Returns:
-            Beautifully formatted tool output with visual enhancements
-        """
-        logger.info(f"🎨 Formatting output for {tool_name}")
-
-        data = {
-            "tool": tool_name,
-            "output": output,
-            "success": success
-        }
-
-        result = hexstrike_client.safe_post("api/visual/tool-output", data)
-        if result.get("success"):
-            logger.info(f"✅ Tool output formatted successfully for {tool_name}")
-        else:
-            logger.error(f"❌ Failed to format tool output for {tool_name}")
-
-        return result
-
-    @mcp.tool()
-    def create_scan_summary(target: str, tools_used: str, vulnerabilities_found: int = 0,
-                           execution_time: float = 0.0, findings: str = "") -> Dict[str, Any]:
-        """
-        Create a comprehensive scan summary report with beautiful visual formatting.
-
-        Args:
-            target: Target that was scanned
-            tools_used: Comma-separated list of tools used
-            vulnerabilities_found: Number of vulnerabilities discovered
-            execution_time: Total execution time in seconds
-            findings: Additional findings or notes
-
-        Returns:
-            Beautiful scan summary report with visual enhancements
-        """
-        logger.info(f"📊 Creating scan summary for {target}")
-
-        tools_list = [tool.strip() for tool in tools_used.split(",")]
-
-        summary_data = {
-            "target": target,
-            "tools_used": tools_list,
-            "execution_time": execution_time,
-            "vulnerabilities": [{"severity": "info"}] * vulnerabilities_found,  # Mock data for count
-            "findings": findings
-        }
-
-        result = hexstrike_client.safe_post("api/visual/summary-report", summary_data)
-        if result.get("success"):
-            logger.info("✅ Scan summary created successfully")
-        else:
-            logger.error("❌ Failed to create scan summary")
-
-        return result
-
-    @mcp.tool()
-    def display_system_metrics() -> Dict[str, Any]:
-        """
-        Display current system metrics and performance indicators with visual formatting.
-
-        Returns:
-            System metrics with beautiful visual presentation
-        """
-        logger.info("📈 Fetching system metrics")
-
-        # Get telemetry data
-        telemetry_result = hexstrike_client.safe_get("api/telemetry")
-
-        if telemetry_result.get("success", True):
-            logger.info("✅ System metrics retrieved successfully")
-
-            # Format the metrics for better display
-            metrics = telemetry_result.get("system_metrics", {})
-            stats = {
-                "cpu_percent": metrics.get("cpu_percent", 0),
-                "memory_percent": metrics.get("memory_percent", 0),
-                "disk_usage": metrics.get("disk_usage", 0),
-                "uptime_seconds": telemetry_result.get("uptime_seconds", 0),
-                "commands_executed": telemetry_result.get("commands_executed", 0),
-                "success_rate": telemetry_result.get("success_rate", "0%")
-            }
-
-            return {
-                "success": True,
-                "metrics": stats,
-                "formatted_display": f"""
-🖥️  System Performance Metrics:
-├─ CPU Usage: {stats['cpu_percent']:.1f}%
-├─ Memory Usage: {stats['memory_percent']:.1f}%
-├─ Disk Usage: {stats['disk_usage']:.1f}%
-├─ Uptime: {stats['uptime_seconds']:.0f}s
-├─ Commands Executed: {stats['commands_executed']}
-└─ Success Rate: {stats['success_rate']}
-""",
-                "timestamp": telemetry_result.get("timestamp", "")
-            }
-        else:
-            logger.error("❌ Failed to retrieve system metrics")
-            return telemetry_result
-
-    # ============================================================================
-    # INTELLIGENT DECISION ENGINE TOOLS
-    # ============================================================================
-
-    @mcp.tool()
-    def analyze_target_intelligence(target: str) -> Dict[str, Any]:
-        """
-        Analyze target using AI-powered intelligence to create comprehensive profile.
-
-        Args:
-            target: Target URL, IP address, or domain to analyze
-
-        Returns:
-            Comprehensive target profile with technology detection, risk assessment, and recommendations
-        """
-        logger.info(f"🧠 Analyzing target intelligence for: {target}")
-
-        data = {"target": target}
-        result = hexstrike_client.safe_post("api/intelligence/analyze-target", data)
-
-        if result.get("success"):
-            profile = result.get("target_profile", {})
-            logger.info(f"✅ Target analysis completed - Type: {profile.get('target_type')}, Risk: {profile.get('risk_level')}")
-        else:
-            logger.error(f"❌ Target analysis failed for {target}")
-
-        return result
-
-    @mcp.tool()
-    def select_optimal_tools_ai(target: str, objective: str = "comprehensive") -> Dict[str, Any]:
-        """
-        Use AI to select optimal security tools based on target analysis and testing objective.
-
-        Args:
-            target: Target to analyze
-            objective: Testing objective - "comprehensive", "quick", or "stealth"
-
-        Returns:
-            AI-selected optimal tools with effectiveness ratings and target profile
-        """
-        logger.info(f"🎯 Selecting optimal tools for {target} with objective: {objective}")
-
-        data = {
-            "target": target,
-            "objective": objective
-        }
-        result = hexstrike_client.safe_post("api/intelligence/select-tools", data)
-
-        if result.get("success"):
-            tools = result.get("selected_tools", [])
-            logger.info(f"✅ AI selected {len(tools)} optimal tools: {', '.join(tools[:3])}{'...' if len(tools) > 3 else ''}")
-        else:
-            logger.error(f"❌ Tool selection failed for {target}")
-
-        return result
-
-    @mcp.tool()
-    def optimize_tool_parameters_ai(target: str, tool: str, context: str = "{}") -> Dict[str, Any]:
-        """
-        Use AI to optimize tool parameters based on target profile and context.
-
-        Args:
-            target: Target to test
-            tool: Security tool to optimize
-            context: JSON string with additional context (stealth, aggressive, etc.)
-
-        Returns:
-            AI-optimized parameters for maximum effectiveness
-        """
-        import json
-
-        logger.info(f"⚙️  Optimizing parameters for {tool} against {target}")
-
-        try:
-            context_dict = json.loads(context) if context != "{}" else {}
-        except:
-            context_dict = {}
-
-        data = {
-            "target": target,
-            "tool": tool,
-            "context": context_dict
-        }
-        result = hexstrike_client.safe_post("api/intelligence/optimize-parameters", data)
-
-        if result.get("success"):
-            params = result.get("optimized_parameters", {})
-            logger.info(f"✅ Parameters optimized for {tool} - {len(params)} parameters configured")
-        else:
-            logger.error(f"❌ Parameter optimization failed for {tool}")
-
-        return result
-
-    @mcp.tool()
-    def create_attack_chain_ai(target: str, objective: str = "comprehensive") -> Dict[str, Any]:
-        """
-        Create an intelligent attack chain using AI-driven tool sequencing and optimization.
-
-        Args:
-            target: Target for the attack chain
-            objective: Attack objective - "comprehensive", "quick", or "stealth"
-
-        Returns:
-            AI-generated attack chain with success probability and time estimates
-        """
-        logger.info(f"⚔️  Creating AI-driven attack chain for {target}")
-
-        data = {
-            "target": target,
-            "objective": objective
-        }
-        result = hexstrike_client.safe_post("api/intelligence/create-attack-chain", data)
-
-        if result.get("success"):
-            chain = result.get("attack_chain", {})
-            steps = len(chain.get("steps", []))
-            success_prob = chain.get("success_probability", 0)
-            estimated_time = chain.get("estimated_time", 0)
-
-            logger.info(f"✅ Attack chain created - {steps} steps, {success_prob:.2f} success probability, ~{estimated_time}s")
-        else:
-            logger.error(f"❌ Attack chain creation failed for {target}")
-
-        return result
-
-    @mcp.tool()
-    def intelligent_smart_scan(target: str, objective: str = "comprehensive", max_tools: int = 5) -> Dict[str, Any]:
-        """
-        Execute an intelligent scan using AI-driven tool selection and parameter optimization.
-
-        Args:
-            target: Target to scan
-            objective: Scanning objective - "comprehensive", "quick", or "stealth"
-            max_tools: Maximum number of tools to use
-
-        Returns:
-            Results from AI-optimized scanning with tool execution summary
-        """
-        logger.info(f"{HexStrikeColors.FIRE_RED}🚀 Starting intelligent smart scan for {target}{HexStrikeColors.RESET}")
-
-        data = {
-            "target": target,
-            "objective": objective,
-            "max_tools": max_tools
-        }
-        result = hexstrike_client.safe_post("api/intelligence/smart-scan", data)
-
-        if result.get("success"):
-            scan_results = result.get("scan_results", {})
-            tools_executed = scan_results.get("tools_executed", [])
-            execution_summary = scan_results.get("execution_summary", {})
-
-            # Enhanced logging with detailed results
-            logger.info(f"{HexStrikeColors.SUCCESS}✅ Intelligent scan completed for {target}{HexStrikeColors.RESET}")
-            logger.info(f"{HexStrikeColors.CYBER_ORANGE}📊 Execution Summary:{HexStrikeColors.RESET}")
-            logger.info(f"   • Tools executed: {execution_summary.get('successful_tools', 0)}/{execution_summary.get('total_tools', 0)}")
-            logger.info(f"   • Success rate: {execution_summary.get('success_rate', 0):.1f}%")
-            logger.info(f"   • Total vulnerabilities: {scan_results.get('total_vulnerabilities', 0)}")
-            logger.info(f"   • Execution time: {execution_summary.get('total_execution_time', 0):.2f}s")
-
-            # Log successful tools
-            successful_tools = [t['tool'] for t in tools_executed if t.get('success')]
-            if successful_tools:
-                logger.info(f"{HexStrikeColors.HIGHLIGHT_GREEN} Successful tools: {', '.join(successful_tools)} {HexStrikeColors.RESET}")
-
-            # Log failed tools
-            failed_tools = [t['tool'] for t in tools_executed if not t.get('success')]
-            if failed_tools:
-                logger.warning(f"{HexStrikeColors.HIGHLIGHT_RED} Failed tools: {', '.join(failed_tools)} {HexStrikeColors.RESET}")
-
-            # Log vulnerabilities found
-            if scan_results.get('total_vulnerabilities', 0) > 0:
-                logger.warning(f"{HexStrikeColors.VULN_HIGH}🚨 {scan_results['total_vulnerabilities']} vulnerabilities detected!{HexStrikeColors.RESET}")
-        else:
-            logger.error(f"{HexStrikeColors.ERROR}❌ Intelligent scan failed for {target}: {result.get('error', 'Unknown error')}{HexStrikeColors.RESET}")
-
-        return result
-
-    @mcp.tool()
-    def detect_technologies_ai(target: str) -> Dict[str, Any]:
-        """
-        Use AI to detect technologies and provide technology-specific testing recommendations.
-
-        Args:
-            target: Target to analyze for technology detection
-
-        Returns:
-            Detected technologies with AI-generated testing recommendations
-        """
-        logger.info(f"🔍 Detecting technologies for {target}")
-
-        data = {"target": target}
-        result = hexstrike_client.safe_post("api/intelligence/technology-detection", data)
-
-        if result.get("success"):
-            technologies = result.get("detected_technologies", [])
-            cms = result.get("cms_type")
-            recommendations = result.get("technology_recommendations", {})
-
-            tech_info = f"Technologies: {', '.join(technologies)}"
-            if cms:
-                tech_info += f", CMS: {cms}"
-
-            logger.info(f"✅ Technology detection completed - {tech_info}")
-            logger.info(f"📋 Generated {len(recommendations)} technology-specific recommendations")
-        else:
-            logger.error(f"❌ Technology detection failed for {target}")
-
-        return result
-
-    @mcp.tool()
-    def ai_reconnaissance_workflow(target: str, depth: str = "standard") -> Dict[str, Any]:
-        """
-        Execute AI-driven reconnaissance workflow with intelligent tool chaining.
-
-        Args:
-            target: Target for reconnaissance
-            depth: Reconnaissance depth - "surface", "standard", or "deep"
-
-        Returns:
-            Comprehensive reconnaissance results with AI-driven insights
-        """
-        logger.info(f"🕵️  Starting AI reconnaissance workflow for {target} (depth: {depth})")
-
-        # First analyze the target
-        analysis_result = hexstrike_client.safe_post("api/intelligence/analyze-target", {"target": target})
-
-        if not analysis_result.get("success"):
-            return analysis_result
-
-        # Create attack chain for reconnaissance
-        objective = "comprehensive" if depth == "deep" else "quick" if depth == "surface" else "comprehensive"
-        chain_result = hexstrike_client.safe_post("api/intelligence/create-attack-chain", {
-            "target": target,
-            "objective": objective
-        })
-
-        if not chain_result.get("success"):
-            return chain_result
-
-        # Execute the reconnaissance
-        scan_result = hexstrike_client.safe_post("api/intelligence/smart-scan", {
-            "target": target,
-            "objective": objective,
-            "max_tools": 8 if depth == "deep" else 3 if depth == "surface" else 5
-        })
-
-        logger.info(f"✅ AI reconnaissance workflow completed for {target}")
-
-        return {
-            "success": True,
-            "target": target,
-            "depth": depth,
-            "target_analysis": analysis_result.get("target_profile", {}),
-            "attack_chain": chain_result.get("attack_chain", {}),
-            "scan_results": scan_result.get("scan_results", {}),
-            "timestamp": datetime.now().isoformat()
-        }
-
-    @mcp.tool()
-    def ai_vulnerability_assessment(target: str, focus_areas: str = "all") -> Dict[str, Any]:
-        """
-        Perform AI-driven vulnerability assessment with intelligent prioritization.
-
-        Args:
-            target: Target for vulnerability assessment
-            focus_areas: Comma-separated focus areas - "web", "network", "api", "all"
-
-        Returns:
-            Prioritized vulnerability assessment results with AI insights
-        """
-        logger.info(f"🔬 Starting AI vulnerability assessment for {target}")
-
-        # Analyze target first
-        analysis_result = hexstrike_client.safe_post("api/intelligence/analyze-target", {"target": target})
-
-        if not analysis_result.get("success"):
-            return analysis_result
-
-        profile = analysis_result.get("target_profile", {})
-        target_type = profile.get("target_type", "unknown")
-
-        # Select tools based on focus areas and target type
-        if focus_areas == "all":
-            objective = "comprehensive"
-        elif "web" in focus_areas and target_type == "web_application":
-            objective = "comprehensive"
-        elif "network" in focus_areas and target_type == "network_host":
-            objective = "comprehensive"
-        else:
-            objective = "quick"
-
-        # Execute vulnerability assessment
-        scan_result = hexstrike_client.safe_post("api/intelligence/smart-scan", {
-            "target": target,
-            "objective": objective,
-            "max_tools": 6
-        })
-
-        logger.info(f"✅ AI vulnerability assessment completed for {target}")
-
-        return {
-            "success": True,
-            "target": target,
-            "focus_areas": focus_areas,
-            "target_analysis": profile,
-            "vulnerability_scan": scan_result.get("scan_results", {}),
-            "risk_assessment": {
-                "risk_level": profile.get("risk_level", "unknown"),
-                "attack_surface_score": profile.get("attack_surface_score", 0),
-                "confidence_score": profile.get("confidence_score", 0)
-            },
-            "timestamp": datetime.now().isoformat()
-        }
-
+    # Register vulnerability intelligence tools for CVE analysis, exploit availability, and risk assessment
+    from mcp_tools.vulnerability_intelligence import register_vulnerability_intelligence_tools
+    register_vulnerability_intelligence_tools(mcp, hexstrike_client, logger)
+
+    # Register enhanced visual output tools for better result presentation
+    from mcp_tools.visual_output_tools import register_visual_output_tools
+    register_visual_output_tools(mcp, hexstrike_client, logger)
+    
+    # Register intelligent decision engine tools for AI-powered analysis and recommendations
+    from mcp_tools.intelligent_decision_engine import register_intelligent_decision_engine_tools
+    register_intelligent_decision_engine_tools(mcp, hexstrike_client, logger, HexStrikeColors)
+    
     # ============================================================================
     # BUG BOUNTY HUNTING SPECIALIZED WORKFLOWS
     # ============================================================================
